dotfiles sync

+++ installer/mac-setup.sh
@@ -24,20 +24,52 @@ fi
 echo 'Getting brew packages...'
 brew bundle
 
-echo 'Starting services...'
-sudo brew services start sudo-touchid
-sudo-touchid
+if ! command -v sudo-touchid &>/dev/null; then
+  echo 'Installing Sudo TouchID...'
+  curl -# https://raw.githubusercontent.com/artginzburg/sudo-touchid/main/sudo-touchid.sh -o /usr/local/bin/sudo-touchid &&
+    chmod +x /usr/local/bin/sudo-touchid &&
+    sudo curl -# https://raw.githubusercontent.com/artginzburg/sudo-touchid/main/com.user.sudo-touchid.plist -o /Library/LaunchDaemons/com.user.sudo-touchid.plist &&
+    /usr/local/bin/sudo-touchid
+fi
+
+# TODO: this requires XCode to be installed, maybe it can be added via `mas`
+# CLI?
+# NOTE: check this page if getting an error from xcodebuild:
+# https://stackoverflow.com/questions/17980759/xcode-select-active-developer-directory-error
+if ! command -v unicornleap &>/dev/null; then
+  echo 'Installing unicornleap... (requires XCode)'
+  mkdir -p ~/dev/forks
+
+  # clone unicornleap
+  rm -rf ~/dev/forks/unicornleap
+  git clone --depth=1 git@github.com:jgdavey/unicornleap.git ~/dev/forks/unicornleap
+
+  # create install dir
+  mkdir ~/.bin
+
+  # Install
+  # shellcheck disable=2164
+  pushd ~/dev/forks/unicornleap &&
+    make &&
+    make images &&
+    cp build/unicornleap ~/.bin/ &&
+    popd || exit
+fi
 
 # Install TerminalVim
-echo 'Installing TerminalVim app'
-cp -r ./iterm/TerminalVim.app /Applications/TerminalVim.app
+if [[ ! -f "/Applications/TerminalVim.app" ]]; then
+  echo 'Installing TerminalVim app'
+  cp -r ./iterm/TerminalVim.app /Applications/TerminalVim.app
+fi
 
-echo 'Installing BTop app'
-cp -r ./iterm/BTop.app /Applications/BTop.app
+if [[ ! -f "/Applications/BTop.app" ]]; then
+  echo 'Installing BTop app'
+  cp -r ./iterm/BTop.app /Applications/BTop.app
+fi
 
 # setup file handlers for TerminalVim
 terminal_vim_id="$(osascript -e 'id of app "TerminalVim"')"
-for ext in md txt js ts json lua rb ex exs eex heex; do
+for ext in md txt js jsx ts tsx json lua rb ex exs eex heex yaml yml plist; do
   echo "Setting TerminalVim as handler for .$ext"
-  duti -s "$terminal_vim_id" ".$ext" all
+  duti -s "$terminal_vim_id" ".$ext" editor
 done
